python nfpm: Split billing and core network db mgmt
This commit splits the functionality for managing the billing database
and managing the core network's user database into separate files
which are installed by different sub-packages.
