+++ nfpm-colte-cn-4g.yaml
@@ -89,6 +89,9 @@ contents:
     file_info:
       mode: 0o444
 
+  - src: "./python/coltedb_cn_4g.py"
+    dst: "/usr/lib/python3/dist-packages/coltedb_cn_4g.py"
+
 # Empty folders your package may need created. (overridable)
 empty_folders:
